+++ dart/test/hub_test.dart
@@ -70,13 +70,18 @@ void main() {
     test('should capture exception', () {
       hub.captureException(fakeException);
 
-      verify(client.captureException(fakeException)).called(1);
+      verify(client.captureException(fakeException, scope: anyNamed('scope')))
+          .called(1);
     });
 
     test('should capture message', () {
       hub.captureMessage(fakeMessage.formatted, level: SentryLevel.info);
       verify(
-        client.captureMessage(fakeMessage.formatted, level: SentryLevel.info),
+        client.captureMessage(
+          fakeMessage.formatted,
+          level: SentryLevel.info,
+          scope: anyNamed('scope'),
+        ),
       ).called(1);
     });
   });